add spotify api call in spotifyService
