+++ src/rdx/actions/types.js
@@ -2,4 +2,4 @@ export const LOGGED_IN = "LOGGED_IN";
 export const COMPLETED_BLOCKER = "COMPLETED_BLOCKER";
 export const COMPLETED_SUGGESTION = "COMPLETED_SUGGESTION";
 export const RESET_PROGRESS = "RESET_PROGRESS";
-export const GET_ACCESS_TOKEN = "GET_ACCESS_TOKEN";
+export const GET_ACCESS_TOKEN_WATCHER = "GET_ACCESS_TOKEN_WATCHER";
